# internal/client/gen/api.yaml
openapi: 3.0.0
info:
  title: OSAPI - A CRUD API for managing Linux systems
  description: |
    OSAPI is a lightweight, minimalistic RESTful API designed to manage and
    configure Linux systems as appliances. The API provides comprehensive
    management capabilities, including reading the current state of the system
    and making privileged, asynchronous changes to match a desired configuration
    state.

    OSAPI is built with pluggability and extensibility in mind, allowing it to
    adapt to different Linux distributions and environments by implementing
    custom handlers for various system operations.

    Key Features:
    - **RESTful API Design:** A RESTful approach for managing Linux systems,
      enabling clear and organized endpoints for system configurations,
      diagnostics, and state management.
    - **Pluggable Architecture:** Easily extend or replace components to suit
      different Linux environments. This flexibility allows users to customize
      behavior based on specific distribution requirements or operational needs.
    - **Job System for Privileged Actions:** Asynchronous job execution via
      NATS JetStream to perform system modifications securely, ensuring that
      changes are managed in a controlled and orderly manner.
    - **System Information Gathering:** Provides detailed system insights, such
      as OS details, hostname, and other environment-specific information,
      facilitating diagnostics and monitoring.
...
  contact:
...
paths:
...
  /system/status:
    servers: []
    get:
      summary: Retrieve system status
      description: >-
        Get the current status of the system including hostname, uptime, load
        averages, memory, and disk usage.
      tags:
        - System_Management_API_system_status
      security:
        - BearerAuth:
            - read
      parameters:
        - name: target_hostname
          in: query
          required: false
          schema:
            type: string
            default: _any
          description: >
            Target: _any (load-balanced), _all (broadcast), hostname (direct),
            or key:value (label group, e.g., group:web.dev).
      responses:
        '200':
          description: A JSON object containing the system's status information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatusCollectionResponse'
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error retrieving system status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /system/hostname:
    servers: []
    get:
      summary: Retrieve system hostname
      description: Get the current hostname of the system.
      tags:
        - System_Management_API_system_operations
      security:
        - BearerAuth:
            - read
      parameters:
        - name: target_hostname
          in: query
          required: false
          schema:
            type: string
            default: _any
          description: >
            Target: _any (load-balanced), _all (broadcast), hostname (direct),
            or key:value (label group, e.g., group:web.dev).
      responses:
        '200':
          description: A JSON object containing the system's hostname.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostnameCollectionResponse'
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error retrieving hostname.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error that occurred.
          example: Failed to retrieve status.
        details:
          type: string
          description: Additional details about the error.
          example: Failed due to network timeout.
        code:
          type: integer
          description: The error code.
          example: 500
...
    LoadAverageResponse:
      type: object
      description: The system load averages for 1, 5, and 15 minutes.
      properties:
        1min:
          type: number
          description: Load average for the last 1 minute.
          example: 0.32
        5min:
          type: number
          description: Load average for the last 5 minutes.
          example: 0.28
        15min:
          type: number
          description: Load average for the last 15 minutes.
          example: 0.25
      required:
        - 1min
        - 5min
        - 15min
    MemoryResponse:
      type: object
      description: Memory usage information.
      properties:
        total:
          type: integer
          description: Total memory in bytes.
          example: 8388608
        free:
          type: integer
          description: Free memory in bytes.
          example: 2097152
        used:
          type: integer
          description: Used memory in bytes.
          example: 4194304
      required:
        - total
        - free
        - used
    DiskResponse:
      type: object
      description: Local disk usage information.
      properties:
        name:
          type: string
          description: Disk identifier, e.g., "/dev/sda1".
          example: /dev/sda1
        total:
          type: integer
          description: Total disk space in bytes.
          example: 500000000000
        used:
          type: integer
          description: Used disk space in bytes.
          example: 250000000000
        free:
          type: integer
          description: Free disk space in bytes.
          example: 250000000000
      required:
        - name
        - total
        - used
        - free
    DisksResponse:
      type: array
      description: List of local disk usage information.
      items:
        $ref: '#/components/schemas/DiskResponse'
    SystemStatusResponse:
      type: object
      properties:
        hostname:
          type: string
          description: The hostname of the system.
          example: my-linux-server
        uptime:
          type: string
          description: The uptime of the system.
          example: 0 days, 4 hours, 1 minute
        load_average:
          $ref: '#/components/schemas/LoadAverageResponse'
        memory:
          $ref: '#/components/schemas/MemoryResponse'
        disks:
          $ref: '#/components/schemas/DisksResponse'
        os_info:
          $ref: '#/components/schemas/OSInfoResponse'
      required:
        - hostname
        - uptime
        - load_average
        - memory
        - disks
        - os_info
    HostnameResponse:
      type: object
      description: The hostname of the system.
      properties:
        hostname:
          type: string
          description: The system's hostname.
          example: my-linux-server
      required:
        - hostname
    SystemStatusCollectionResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          description: The job ID used to process this request.
          example: 550e8400-e29b-41d4-a716-446655440000
        results:
          type: array
          items:
            $ref: '#/components/schemas/SystemStatusResponse'
      required:
        - results
    HostnameCollectionResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          description: The job ID used to process this request.
          example: 550e8400-e29b-41d4-a716-446655440000
        results:
          type: array
          items:
            $ref: '#/components/schemas/HostnameResponse'
      required:
        - results
    OSInfoResponse:
      type: object
      description: Operating system information.
      properties:
        distribution:
          type: string
          description: The name of the Linux distribution.
          example: Ubuntu
        version:
          type: string
          description: The version of the Linux distribution.
          example: '20.04'
      required:
        - distribution
        - version
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...